removes target/

+++ src/main/java/pt/unl/fct/pds/utils/CandidateNode.java
@@ -4,16 +4,16 @@ import pt.unl.fct.pds.model.Node;
 
 public class CandidateNode {
     private Node node;
-    private int weightedBandwidth;
+    private double weightedBandwidth;
 
-    public CandidateNode(Node node, int weightedBandwidth) {
+    public CandidateNode(Node node, double weightedBandwidth) {
         this.node = node;
         this.weightedBandwidth = weightedBandwidth;
     }
 
     public Node getNode() { return node; }
-    public int getWeightedBandwidth() { return weightedBandwidth; }
+    public double getWeightedBandwidth() { return weightedBandwidth; }
 
     public void setNode(Node node) { this.node = node; }
-    public void setWeightedBandwidth(int weightedBandwidth) { this.weightedBandwidth = weightedBandwidth; }
+    public void setWeightedBandwidth(double weightedBandwidth) { this.weightedBandwidth = weightedBandwidth; }
 }
